Convert setting between JSON and value in form
Makes the form easier to work with, dont have to give it
any extra values or parse the cleaned data in a special way.
Converting from {"email_address": "me@mail.com"} to "me@mail.com"
and opposite way is now done in the form. Feels a bit bad
since I just straight up replace the old settings field with
a new CharField, so some validation might go lost, but I dont
think the JSONField validated anything outside being valid JSON,
so its probably fine

A sizeable commit, since a lot of stuff had to be changed
for this to work. Media had to be included in the
update form so the settings_key could be found,
but then the template had to be updated to include hidden fields
but not actually show them